+++ HISTORY.rst
@@ -100,6 +100,8 @@
 * Backport sidebar code.
 * CSS updates to maximize usable width.
 * Improve partial navigation error messages and display.
+* Add user created and user edited events.
+* Add group created and group edited events.
 
 3.1.11 (2019-04-XX)
 ===================